% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/importPeakFQ.R
\name{importPeakFQ}
\alias{importPeakFQ}
\title{Import Data from PeakFQ Output}
\usage{
importPeakFQ(pfqPath, gisFile, sites = "")
}
\arguments{
\item{pfqPath}{A directory that contains all of PeakFQ files for each site
in the GIS file.  Files can be contained in the main path or in
subdirectories of \code{pfqPath}.  Each site should be represented by one
 and only one .EXP and .PRT file.}

\item{gisFile}{A tab-delimited text file that contains a matrix whose rows
represent the sites and contains columns that include \sQuote{Station.ID},
\sQuote{Lat}, \sQuote{Long} and any other variables to be used for analysis.}

\item{sites}{(optional) A vector containing the independent variables (as above) for a
particular target site.  This variable is only used for ROI analysis.}
}
\value{
All outputs are returned as part of a list.  The list includes:
\item{sites}{A vector of site IDs.}
\item{Y}{A matrix whose comlumns represent unique frequency events,
while the row represent particular sites in the same order as \code{sites}.}
\item{AEP}{A vector of annual exceedance probabilities associated with the
columns of \code{Y}.}
\item{X}{A matrix whose columns represent basin characteristics to be used
 as dependent variables and whose rows represent sites corresponding to \code{sites}.}
\item{LP3f}{A matrix containing the fitted LP# parameters that are fixed
across exceedence probability.  These include the standard deviation,
skew and regional skew for each site.}
\item{LP3k}{A matrix of the fitted kappa parameters of the LP3 distribution
 for each \code{AEP}.}
\item{BasChars}{A matrix containing the site IDs, latitudes and longitudes.}
\item{MSEGR}{The mean-squared-error in the regional skew.}
\item{recLen}{A square matrix indicating the number of overlapping years
 for each site pair.}
\item{recCor}{A matrix of the correlaiton between site paris.}
}
\description{
The \code{importPeakFQ} function reads output directly from PeakFQ and
prepares it for use in WREG-R.
}
\details{
This functions allows users to read output directly from PeakFQ without
the need to manipulate additional files.  The site selection is driven
by the sites in the GIS table.
}

