% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPeakFQ_oneFile.R
\name{importPeakFQ_oneFile}
\alias{importPeakFQ_oneFile}
\title{Import Data from PeakFQ Output}
\usage{
importPeakFQ_oneFile(pfqPath, destination = "peakFQ_extract")
}
\arguments{
\item{pfqPath}{A directory that contains a single EXP file and a single PRT 
file, both from peakFQ and both containing the same gage IDs.}

\item{destination}{The directory name that will be created to hold all 
of the output from this function.}
}
\value{
Creates a directory structure in \code{destination} that contains a 
 directory for each station number in the EXP file. Each gage subdirectory
 will contain the resulting single-site EXp and PRT files.
}
\description{
The \code{importPeakFQ_oneFile} function reads output from PeakFQ when the 
output is contained in a a single EXP file and a single PRT file.
}
\details{
This function allows users to parse the output from peakFQ when all 
sites are contained in a single EXP and a single PRT file. This function 
will create new files, with a single EXP and a single PRT for each site, 
that can then be imported using \code{\link{importPeakFQ}}.

WARNING: This function will create new files on your computer.
}
\examples{
peakFQdir <- paste0(
  file.path(system.file("exampleDirectory", package = "WREG"),
    "pfqImport_oneFile"))
importedData <- importPeakFQ_oneFile(pfqPath = peakFQdir)

}
