% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{Influence}
\alias{Influence}
\title{Influence Statistics (Cook's D) (WREG)}
\usage{
Influence(e, X, Omega, Beta, ROI = FALSE, Lev = NA)
}
\arguments{
\item{e}{contains the model residuals.}

\item{X}{contains independent variables.  A leading constant of one is included if 
a constant is included in the regression.  Each row represents a unique obsevation.}

\item{Omega}{is the weighting matrix used for regression fitting.}

\item{Beta}{contains the fitted model coefficients.}

\item{ROI}{is a logical indicating if this is a region-of-influence regression.}

\item{Lev}{is a vector with the same length as \code{e} and includes the leverage of 
each observation.  
This input is required for any region-of-influence regression.}
}
\value{
The function returns as list as output.  The list contains:
\item{Influence}{A vector containing the influence (Cook's D) of each observation 
on the estimated regression coefficients.}
\item{Limit}{The critical influence value for this dataset.}
\item{Significant}{A logical vector the same size as \code{Influence}. 
 It indicates if the influence is significant for each observation.}
}
\description{
The \code{Influence} function calculates the influence statistics (Cooks-D) for any 
regrerssion.
}
\details{
Influence is a measure of the impact each observation has on the estimated 
regression coefficients.  The calculation is based on equation 43 of the WREG 
v. 1.0 manual.  The critical value of influence is calculated using equation 44.  
An influence is considered significant if the absolute value of the influence 
is greater than the critical value.

For region-of-influence regressions, the influence calculation is weighted 
by the leverage of that observation on the target site and the overall leverage 
of the observation.  This is a departure from the WREG v. 1.0, but reflects the 
WREG v. 1.05 code.
}

