% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leverage.R
\name{Leverage}
\alias{Leverage}
\title{Leverage Statistics (WREG)}
\usage{
Leverage(X, Omega, Ch = NA, x0 = NA, ROI = FALSE)
}
\arguments{
\item{X}{contains independent variables.  A leading constant of one is
included if a constant is included in the regression.  Each row represents a
unique observation.}

\item{Omega}{is the weighting matrix used for regression fitting}

\item{Ch}{allows the user to specify a custom criteria for testing.  If not
specified, the default is 4 for region-of-influence regression and 2
elsewise.}

\item{x0}{is used only in the case of region-of-influence regression.  It
contains the independent variables of a specific site against which to
calculate the leverage.}

\item{ROI}{is a logical vector specifying if the regression is or is not 
region-of-influence regression}
}
\value{
The function returns a list as output.  This list contains: 
 \item{Leverage}{A vector indicating the leverage of each observation. This
 is a vector whose length equals the number of rows in X.} \item{Limit}{A
 number indicating the critical value of leverage for this data set.} 
 \item{Significant}{A logical vector the same size as \code{Leverage}. It
 indicates if the leverage is significant for each observation.}
}
\description{
The \code{Leverage} function calculates the leverage statistics
for each observation in the regression.
}
\details{
Leverage is a measure of how far way the independent variables of an
observation are from the other observations in the regression set.  A leverage
is considered significant if the absolute value of the leverage is greater
than the critical value.  These calculations are based on equations 40, 41 and
42 of the WREG v. 1.0 manual.
}
\examples{
# Import some example data
peakFQdir <- paste0(
  file.path(system.file("exampleDirectory", package = "WREG"),
    "pfqImport"))
gisFilePath <- file.path(peakFQdir, "pfqSiteInfo.txt")
importedData <- importPeakFQ(pfqPath = peakFQdir, gisFile = gisFilePath)

# Run a simple regression
Y <- importedData$Y$AEP_0.5
X <- importedData$X[c("Sand", "OutletElev", "Slope")]
transY <- "none"
result <- WREG.OLS(Y, X, transY)

# calculate leverage of each point
leverageResult <- Leverage(X = X, 
  Omega = result$Weighting)

}
