% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.OLS}
\alias{print.WREG.GLS}
\alias{print.WREG.GLSs}
\alias{print.WREG.OLS}
\alias{print.WREG.WLS}
\title{print.WREG}
\usage{
\method{print}{WREG.OLS}(object)

\method{print}{WREG.WLS}(object)

\method{print}{WREG.GLS}(object)

\method{print}{WREG.GLSs}(object)
}
\arguments{
\item{object}{An output list from one of the WREG... functions}
}
\value{
\code{print.WREG.OLS} Prints a summary of output list from WREG.OLS

\code{print.WREG.WLS} Prints a summary of output list from WREG.WLS

\code{print.WREG.GLS} Prints a summary of output list from WREG.GLS

\code{print.GLSs} Prints a summary of output list from WREG.GLSs
}
\description{
Print methods for WREG output lists
}
\details{
print.WREG
}
\note{
\code{print} is a generic name for the functions documented.
\cr
If called, \code{print} displays a summary of the output from WREG... functions
}

