% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.WREG}
\alias{print.WREG}
\alias{print.WREG.GLS}
\alias{print.WREG.GLSs}
\alias{print.WREG.OLS}
\alias{print.WREG.WLS}
\title{Print output (WREG)}
\usage{
\method{print}{WREG}(object)
}
\arguments{
\item{object}{is an output from \code{WREG.MLR}.}
}
\description{
The \code{print} function summarizes the regression output from WREG.
}
\details{
This function is intended to replicate the output from the original WREG 
matlab program in a format similar to the generic \code{summary.lm} function.
}

