% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regEquation.R
\name{regEquation}
\alias{regEquation}
\title{Print Resulting Regresstion String}
\usage{
regEquation(object, decimals = 2)
}
\arguments{
\item{object}{The dependent variable of interest, with any transformations 
already applied.}

\item{decimals}{An integer or vector of integers specifying how many decimal 
places should be used for each coefficient. If provided as a vector that 
does not match the number of rows in \code{object$Coefs}, then the first 
number provided will be replicated for all coefficients.}
}
\value{
A character string using the estimated coefficients and the 
 coefficient names from the \code{object}
}
\description{
The \code{regEquation} function takes a WREG object and returns
the equation string.
}
\details{
This function is used via a the print command, but can also be used 
 by itself.
}
\examples{
# Import some example data
peakFQdir <- paste0(
  file.path(system.file("exampleDirectory", package = "WREG"),
    "pfqImport"))
gisFilePath <- file.path(peakFQdir, "pfqSiteInfo.txt")
importedData <- importPeakFQ(pfqPath = peakFQdir, gisFile = gisFilePath)

# Run a simple regression
Y <- importedData$Y$AEP_0.5
X <- importedData$X[c("Sand", "OutletElev", "Slope")]
X <- cbind(data.frame("Intercept" = rep(1, nrow(X))), X)
transY <- "none"
result <- WREG.OLS(Y, X, transY)
print(regEquation(result))

}
