% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regEquationFormat.R
\name{regEquationFormat}
\alias{regEquationFormat}
\title{Utility function to format the regression equation}
\usage{
regEquationFormat(X, C1 = NULL, C2 = NULL, C3 = NULL, C4 = NULL, mode,
  var)
}
\arguments{
\item{X}{The X/Y variable}

\item{C1}{(optional) Constant 1 for transform}

\item{C2}{(optional) Constant 2 for transform}

\item{C3}{(optional) Constant 3 for transform}

\item{C4}{(optional) Constant 4 for transform}

\item{mode}{Type of transform on X/Y variable}

\item{var}{Whether or not method applies to the X or Y variable}
}
\value{
Nothing, strings are added to list xEq
}
\description{
The \code{regEquationFormat} function formats 
the regression equation for WREG-R.
}
\details{
This functions formats the regression equation based on input for WREG output.
}
\examples{
regEquationFormat("P", mode="none", var="X")

}
