% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testXCorr.R
\name{testXCorr}
\alias{testXCorr}
\title{Exploring correlation function}
\usage{
testXCorr(siteData, BasinChars, alpha, theta, DistMeth = c(1, 2),
  concurrentMin = 25, plot = TRUE)
}
\arguments{
\item{siteData}{A list of the records at each site.  Each element of the list,
representing a site, is a data frame with 2 variables names \code{year} 
and \code{obs}.}

\item{BasinChars}{A dataframe containing three variables: \code{StationID} is the 
numerical identifier (without a leading zero) of each site, \code{Lat} is the latitude
of the site, in decimal degrees, and \code{Long} is the longitude of the site, in decimal
degrees.}

\item{alpha}{A number, \code{alpha} is a parameter used in the estimated 
cross-correlation between site records.  See equation 20 in the WREG 
v. 1.05 manual.}

\item{theta}{A number, \code{theta} is a parameter used in the estimated 
cross-correlation between site records.  See equation 20 in the WREG 
v. 1.05 manual.}

\item{DistMeth}{A value of \code{1} indicates that the "Nautical Mile" 
approximation should be used to calculate inter-site distances.  A value 
of \code{2} designates the Haversine approximation.  See 
\code{\link{Dist.WREG}}.}

\item{concurrentMin}{A number specifying the minimum number of years of 
concurrent record required to estimate cross-correlation.}

\item{plot}{A logical spcifying if the plot should be created.}
}
\value{
If \code{plot=FALSE}, the Nash-Sutcliffe model efficieny is returned.
}
\description{
\code{testXCorr} explores the fit of the proposed approximation of 
cross-correlation.
}

