% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn.R
\name{warn}
\alias{warn}
\title{Utility function to add and remove warnings without halting execution}
\usage{
warn(mode, w = NULL)
}
\arguments{
\item{mode}{Whether to initialize, check, get, add, or clear warnings}

\item{w}{(optional) Message for adding warning}
}
\value{
Nothing, hasWarnings, or string list of warnings
}
\description{
The \code{warn} handles warnings
for WREG-R.
}
\details{
This functions is a warning scheme that does not halt execution.
}
\examples{
hasWarnings <- warn("check")

}
