% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WREG.GLS.R
\name{WREG.GLS}
\alias{WREG.GLS}
\title{Weighted-Multiple-Linear Regression Program (WREG)}
\usage{
WREG.GLS(Y, X, recordLengths, LP3, basinChars, transY, x0 = NA,
  alpha = 0.01, theta = 0.98, peak = T, distMeth = 2, regSkew = FALSE,
  MSEGR = NA, TY = 2, legacy = FALSE)
}
\arguments{
\item{Y}{A numeric vector of the dependent variable of interest, with any 
transformations already applied.}

\item{X}{A numeric matrix of the independent variables in the regression, with
any transformations already applied.  Each row represents a site and each 
column represents a particular independe variable.  (If a leading constant 
is used, it should be included here as a leading column of ones.)  The rows 
must be in the same order as the dependent variables in \code{Y}.}

\item{LP3}{A numeric matrix containing the fitted Log-Pearson Type III 
standard deviate, standard deviation and skew for each site.  The columns of
the matrix represent S, K, G, and an option regional skew value \code{GR} 
required by WREG.GLS with regSkew = TRUE. The order of the rows must be the 
same as \code{Y}.}

\item{transY}{A required character string indicating if the the 
dependentvariable was transformed by the common logarithm ('log10'), 
transformed by the natural logarithm ('ln') or untransformed ('none').}

\item{x0}{A vector containing the independent variables (as above) for a 
particular target site.  This variable is only used for ROI analysis.}

\item{alpha}{A numeric. \code{alpha} is a parameter used in the estimated 
cross-correlation between site records.  See equation 20 in the WREG v. 1.05
manual.  The arbitrary, default value is 0.01.  The user should fit a 
different value as needed.}

\item{theta}{A numeric. \code{theta} is a parameter used in the estimated 
cross-correlation between site records.  See equation 20 in the WREG v. 1.05
manual.  The arbitrary, default value is 0.98.  The user should fit a 
different value as needed.}

\item{regSkew}{A logical vector indicating if regional skews are provided with
an adjustment required for uncertainty therein (\code{TRUE}).  The default 
is \code{FALSE}.}

\item{MSEGR}{A numeric. The mean squared error of the regional skew.  Required
only if \code{regSkew = TRUE}.}

\item{TY}{A numeric.  The return period of the event being modeled.  Required 
only for \dQuote{GLSskew}.  The default value is \code{2}.  (See the 
\code{Legacy} details below.)}

\item{RecordLengths}{A numeric matrix whose rows and columns are in the same 
order as \code{Y}.  Each \code{(r,c)} element represents the length of 
concurrent record between sites \code{r} and \code{c}.  The diagonal 
elements therefore represent each site's full record length.}

\item{BasinChars}{A dataframe containing three variables: \code{StationID} is 
the identifier of each site, \code{Lat} is the latitude of the site, in 
decimal degrees, and \code{Long} is the longitude of the site, in decimal 
degrees.  The sites must be presented in the same order as \code{Y}. 
Required only if \code{regSkew = TRUE}.}

\item{Peak}{A logical.  Indicates if the event being modeled is a peak flow 
event or a low-flow event.  \code{TRUE} indicates a peak flow, while 
\code{FALSE} indicates a low-flow event.}

\item{CustomWeight}{This allows the user to enter a custom weighting matrix. 
It is included also to provide legacy code for WREG v. 1.05. 
\code{CustomWeight} can either be a square matrix of weights with a length 
equal to \code{length(Y)} or a list containing three elements.  The elements
of the list include (1) \code{Omega} as the square weighting matrix, (2) 
\code{var.modelerror.k} as the estimated variance of the model errors from 
the k-variable model (\code{k=ncol(X)}), and (3) \code{var.modelerror.0} as 
the variance of the model errors from a consant-only regression.  Required 
for \code{Reg=} \dQuote{CustomWeight}.}

\item{DistMeth}{A numeric. A value of \code{1} indicates that the "Nautical 
Mile" approximation should be used to calculate inter-site distances.  A 
value of \code{2} designates the Haversine approximation.  See 
\code{\link{Dist.WREG}}.  The default value is \code{2}.  (See the 
\code{Legacy} details below.)}

\item{Legacy}{A logical.  A value of \code{TRUE} forces the WREG program to 
behave identically to WREG v. 1.05, with BUGS and all.  It will force 
\code{TY=2} and \code{DistMeth=1}.  For ROI regressions, it will also 
require a specific calculation for weighing matrices in \dQuote{WLS} 
(\code{\link{Omega.WLS.ROImatchMatLab}}), \dQuote{GLS}, and \dQuote{GLSskew}
(see \code{\link{Omega.GLS.ROImatchMatLab}}) Further details are provided in
\code{\link{WREG.RoI}}}
}
\value{
All outputs are returned as part of a list.  The elements of the list 
 depend on the type of regression performed.  The elements of the list may 
 include: \item{Coefs}{A data frame composed of four variables: (1) 
 \code{Coefficient} contains the regression coefficeints estimated for the 
 model, (2) \code{\sQuote{Standard Error}} contains the standard errors of 
 each regression coefficient, (3) \code{tStatistic} contains the Student's 
 T-statistic of each regression coefficient and (4) \code{pValue} contains 
 the significance probability of each regression coefficient.} 
 \item{ResLevInf}{A data frame composed of three variables for each site in 
 the regression.  \code{Residual} contains the model residuals. 
 \code{Leverage} contains the leverage of each site.  \code{Influence} 
 contains the influence of each site.} \item{LevLim}{The critical value of 
 leverage.  See \code{\link{Leverage}}} \item{InflLim}{The critical value of 
 influence.  See \code{\link{Influence}}} \item{LevInf.Sig}{A logical matrix 
 indicating if the leverage (column 1) is significant and the influence 
 (column 2) is significant for each site in the regression.} 
 \item{PerformanceMetrics}{A list of not more than ten elements.  All 
 regression types return the mean squared error of residuals (\code{MSE}), 
 the coefficient of determination (\code{R2}), the adjusted coefficient of 
 determination (\code{R2_adj}) and the root mean squared error (\code{RMSE}, 
 in percent).  The pseudo coefficient of regression (\code{R2_pseudo}), the 
 average variance of prediction (\code{AVP}), the standard error of 
 prediction (\code{Sp}, in percent), a vector of the individual variances of 
 prediction for each site (\code{VP.PredVar}), the model-error variance 
 (\code{ModErrVar}) and the standardized model error variance 
 (\code{StanModErr}, in percent) are also returned.  Details on the 
 appropriateness and applicability of performance metrics can be found in the
 WREG manual.} \item{X}{The input predictors.} \item{Y}{The input 
 observations.} \item{fitted.values}{A vector of model estimates from the 
 regression model.} \item{residuals}{A vector of model residuals.} 
 \item{Weighting}{The weighting matrix used to develop regression estimates.}
 \item{Input}{A list of input parameters for error searching.  Currently 
 empty.}
}
\description{
The \code{WREG.GLS} function executes the multiple linear 
 regression analysis using generalized least-squares regression.
}
\details{
In this implementation, the weights for generalized least-squares 
 regression are defined by intersite correlations and record lengths.  See
 manual for details.
 
 The logical handle \code{Legacy} has been included to test that this program
 returns the same results as WREG v. 1.05.  In the development of this code, 
 some idiosyncrasies of the MatLab code for WREG v. 1.05 became apparent. 
 Setting \code{Legacy} equal to \code{TRUE} forces the code to use the same 
 idiosycrasies as WREG v. 1.05.  Some of these idosyncrasies may be bugs in 
 the code.  Further analysis is needed.  For information on the specific 
 idiosyncrasies, see the notes for the \code{Legacy} input and the links to 
 other functions in this package.
}
\examples{
\dontrun Add example
}

