% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WREG.RoI.R
\name{WREG.RoI}
\alias{WREG.RoI}
\title{Region-of-Influence Regression (WREG)}
\usage{
WREG.RoI(Y, X, Reg, transY = NA, RecordLengths = NA, LP3 = NA,
  regSkew = FALSE, alpha = 0.01, theta = 0.98, BasinChars = NA,
  MSEGR = NA, TY = 2, Peak = T, ROI = c("PRoI", "GRoI", "HRoI"),
  n = NA, D = 250, DistMeth = 2, Legacy = FALSE)
}
\arguments{
\item{Y}{The dependent variable of interest, with any transformations already 
applied.}

\item{X}{The independent variables in the regression, with any transformations
already applied.  Each row represents a site and each column represents a
particular independe variable.  (If a leading constant is used, it should be
included here as a leading column of ones.)  The rows must be in the same
order as the dependent variables in \code{Y}.}

\item{Reg}{A string indicating which type of regression should be applied. The
options include: \dQuote{OLS} for ordinary least-squares regression, 
\dQuote{WLS} for weighted least-squares regression, \dQuote{GLS} for
generalized least-squares regression, with no uncertainty in regional skew
and \dQuote{GLSskew} for generalized least-squares regression with
uncertainty in regional skew. (In the case of \dQuote{GLSskew}, the
uncertainty in regional skew must be provided as the mean squared error in
regional skew.)}

\item{transY}{A required character string indicating if the the 
dependentvariable was transformed by the common logarithm ('log10'), 
transformed by the natural logarithm ('ln') or untransformed ('none').}

\item{RecordLengths}{This input is required for \dQuote{WLS}, \dQuote{GLS} and
\dQuote{GLSskew}.  For \dQuote{GLS} and \dQuote{GLSskew},
\code{RecordLengths} should be a matrix whose rows and columns are in the
same order as \code{Y}.  Each \code{(r,c)} element represents the length of
concurrent record between sites \code{r} and \code{c}.  The diagonal
elements therefore represent each site's full record length.  For
\dQuote{WLS}, the only the at-site record lengths are needed. In the case of
\dQuote{WLS}, \code{RecordLengths} can be a vector or the matrix described
for \dQuote{GLS} and \dQuote{GLSskew}.}

\item{LP3}{A dataframe containing the fitted Log-Pearson Type III standard 
deviate, standard deviation and skew for each site.  The names of this data
frame are \code{S}, \code{K} and \code{G}.  For \dQuote{GLSskew}, the
regional skew value must also be provided in a variable called \code{GR}. 
The order of the rows must be the same as \code{Y}.}

\item{regSkew}{A logical vector indicating if regional skews are provided with
an adjustment required for uncertainty therein (\code{TRUE}).  The default 
is \code{FALSE}.}

\item{alpha}{A number, required only for \dQuote{GLS} and \dQuote{GLSskew}. 
\code{alpha} is a parameter used in the estimated cross-correlation between
site records.  See equation 20 in the WREG v. 1.05 manual.  The arbitrary,
default value is 0.01.  The user should fit a different value as needed.}

\item{theta}{A number, required only for \dQuote{GLS} and \dQuote{GLSskew}. 
\code{theta} is a parameter used in the estimated cross-correlation between
site records.  See equation 20 in the WREG v. 1.05 manual.  The arbitrary,
default value is 0.98.  The user should fit a different value as needed.}

\item{BasinChars}{A dataframe containing three variables: \code{StationID} is
the numerical identifier (without a leading zero) of each site, \code{Lat}
is the latitude of the site, in decimal degrees, and \code{Long} is the
longitude of the site, in decimal degrees.  The sites must be presented in
the same order as \code{Y}.  Required only for \dQuote{GLS}.}

\item{MSEGR}{A number. The mean squared error of the regional skew.}

\item{TY}{A number.  The return period of the event being modeled.  Required
only for \dQuote{GLSskew}.  The default value is \code{2}.  (See the
\code{Legacy} details below.)}

\item{Peak}{A logical.  Indicates if the event being modeled is a peak flow
event or a low-flow event.  \code{TRUE} indicates a peak flow, while
\code{FALSE} indicates a low-flow event.}

\item{ROI}{A string indicating how to define the region of influence. 
\dQuote{PRoI} signifies physiographic, independent or predictor-variable
region of influence. \dQuote{GRoI} calls for a geographic region of
influence. \dQuote{HRoI} requests a hybrid region of influence.  Details on
each approach are provided in the manual for WREG v. 1.0.}

\item{n}{The number of sites to include in the region of influence.}

\item{D}{Required for \dQuote{HRoI}, the geographic limit within which to
search for a physiographic region of influence.  In WREG v. 1.05 (see
\code{Legacy} below), this is interpretted in meters.  Elsewise, this is
interpretted as miles.}

\item{DistMeth}{Required for \dQuote{GLS} and \dQuote{GLSskew}.  A value of
\code{1} indicates that the "Nautical Mile" approximation should be used to
calculate inter-site distances.  A value of \code{2} designates the
Haversine approximation.  See \code{\link{Dist.WREG}}.  The default value is
\code{2}.  (See the \code{Legacy} details below.)}

\item{Legacy}{A logical.  A value of \code{TRUE} forces the WREG program to
behave identically to WREG v. 1.05, with BUGS and all.  It will force
\code{TY=2} and \code{DistMeth=1}.  For ROI regressions, it will also
require a specific calculation for weighing matrices in \dQuote{WLS}
(\code{\link{Omega.WLS.ROImatchMatLab}}), \dQuote{GLS}, and \dQuote{GLSskew}
(see \code{\link{Omega.GLS.ROImatchMatLab}}). \code{Legacy} also forces the
distance limit \code{D} to be interpretted in meters.}
}
\value{
As with other WREG functions, \code{WREG.RoI} returns a large list
 of regression parameters and metrics.  This list varies depending on the
 \code{Reg} specified, but may contain: \item{fitted.values}{A vector of
 model estimates from the regression model.} \item{residuals}{A vector of
 model residuals.} \item{PerformanceMetrics}{A list of four elements.  These
 represent approximate performance regression across all of the
 region-of-influence regressions. These include the mean squared error of
 residuals (\code{MSE}), the coefficient of determination (\code{R2}), the
 adjusted coefficient of determination (\code{R2_adj}) and the root mean
 squared error (\code{RMSE}, in percent).  Details on the appropriateness and
 applicability of performance metrics can be found in the WREG manual.} 
 \item{Coefficients}{A list composed of four elements: (1) \code{Values}
 contains the regression coefficeints estimated for the model built around
 each observation, (2) \code{StanError} contains the standard errors
 of each regression coefficient for the ROI regressions around each
 observations, (3) \code{TStatistic} contains the Student's T-statistic of
 each regression coefficient for the ROI regression built around each
 observation and (4) \code{pValue} contains the significance probability of
 each regression coefficient for the ROI regressions built around each 
 observation.  Each element of the list is a matrix the same size as
 \code{X}} \item{ROI.Regressions}{A list of elements and outputs from each
 individual ROI regression.  These include a matrix of the sites used in each
 ROI regression (\code{Sites.Used}, a \code{length(Y)}-by-\code{n} matrix), a
 matrix of the geographic distances between the selected sites in each ROI
 regression (\code{ Gdist.Used}, a \code{length(Y)}-by-\code{n} matrix), a
 matrix of the physiographic distances between the selected sites in each ROI
 regression (\code{ Pdist.Used}, a \code{length(Y)}-by-\code{n} matrix), a
 matrix of the observations used in each ROI regression (\code{Obs.Used}, a
 \code{length(Y)}-by- \code{n} matrix), a matrix of model fits in the region
 of influence (\code{Fits}, a \code{length(Y)}-by-\code{n} matrix), a matrix
 of model residuals in the region of influence (\code{Residuals}, a
 \code{length(Y)}-by-\code{n} matrix), a matrix of leverages for each ROI
 regression (\code{Leverage}, a \code{length(Y)}-by-\code{n} matrix), a
 matrix of influences for each ROI regression (\code{Influence}, a \code{ 
 length(Y)}-by-\code{n} matrix), a logical matrix indicating if the leverage
 is significant in each ROI regression (\code{Leverage.Significance}, a
 \code{length(Y)} -by-\code{n} matrix), a logical matrix indicating if the
 influence is significant in each ROI regression
 (\code{Influence.Significance}, a \code{length(Y)} -by-\code{n} matrix), a
 vector of critical leverage values for each ROI regression 
 (\code{Leverage.Limits}), a vector of critical influence values for each ROI
 regression (\code{Influence.Limits}) and a list of performance metrics for
 each ROI regression (\code{PerformanceMetrics}).  The last element, 
 \code{PerformanceMetrics} is identical to the same output from other 
 functions excpet that every element is multiplied by the number of
 observations so as to capture the individual performance of each ROI
 regression.} \item{ROI.InputParameters}{A list of input parameters to record
 the controls on the ROI regression.  \code{D} idicates the limit used in
 \dQuote{HRoI}. \code{n} indicates the size of the region of influence. 
 \code{ROI} is a string indicating the type of region of influence. 
 \code{Legacy} is a logical indicating if the WREG v. 1.05 idiosycrasies were
 implemented.}
}
\description{
\code{WREG.ROI} implements region-of-influence regression in the
WREG framework.
}
\details{
The support for region-of-influence regression is described in the
 manual of WREG v. 1.0.  \code{WREG.RoI} iterates through the sites of
 \code{Y}, defines a region of influence and implements the specified
 regression by calling a WREG function.
 
 The logical handle \code{Legacy} has been included to test that this program
 returns the same results as WREG v. 1.05.  In the development of this code,
 some idiosyncrasies of the MatLab code for WREG v. 1.05 became apparent. 
 Setting \code{Legacy} equal to \code{TRUE} forces the code to use the same
 idiosycrasies as WREG v. 1.05.  Some of these idosyncrasies may be bugs in
 the code.  Further analysis is needed.  For information on the specific
 idiosyncrasies, see the notes for the \code{Legacy} input and the links to
 other functions in this package.
}
\examples{
# Import some example data
peakFQdir <- paste0(
  file.path(system.file("exampleDirectory", package = "WREG"),
    "pfqImport"))
gisFilePath <- file.path(peakFQdir, "pfqSiteInfo.txt")
importedData <- importPeakFQ(pfqPath = peakFQdir, gisFile = gisFilePath)

# Run a simple regression
Y <- importedData$Y$AEP_0.5
X <- importedData$X[c("A")]
transY <- "none"
basinChars <- importedData$BasChars
#result <- WREG.OLS(Y, X, transY)

result <- WREG.RoI(Y, X, Reg = "OLS", transY, BasinChars = basinChars,
  ROI='GRoI', n = 10L)
}

