% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WREG.UW.R
\name{WREG.UW}
\alias{WREG.UW}
\title{Weighted-Multiple-Linear Regression Program (WREG)}
\usage{
WREG.UW(Y, X, customWeight, transY, x0 = NA)
}
\arguments{
\item{Y}{The dependent variable of interest, with any transformations already 
applied.}

\item{X}{The independent variables in the regression, with any transformations
already applied.  Each row represents a site and each column represents a 
particular independe variable.  (If a leading constant is used, it should be
included here as a leading column of ones.)  The rows must be in the same 
order as the dependent variables in \code{Y}.}

\item{customWeight}{This allows the user to enter a custom weighting matrix. 
It is included also to provide legacy code for WREG v. 1.05. 
\code{customWeight} can either be a square matrix of weights with a length 
equal to \code{length(Y)} or a list containing three elements.  The elements
of the list include (1) \code{Omega} as the square weighting matrix, (2) 
\code{var.modelerror.k} as the estimated variance of the model errors from 
the k-variable model (\code{k=ncol(X)}), and (3) \code{var.modelerror.0} as 
the variance of the model errors from a consant-only regression.}

\item{transY}{A required character string indicating if the the 
dependentvariable was transformed by the common logarithm ('log10'), 
transformed by the natural logarithm ('ln') or untransformed ('none').}

\item{x0}{A vector containing the independent variables (as above) for a 
particular target site.  This variable is only used for ROI analysis.}
}
\value{
All outputs are returned as part of a list.  The elements of the list 
 depend on the type of regression performed.  The elements of the list may 
 include: \item{Coefs}{A data frame composed of four variables: (1) 
 \code{Coefficient} contains the regression coefficeints estimated for the 
 model, (2) \code{\sQuote{Standard Error}} contains the standard errors of 
 each regression coefficient, (3) \code{tStatistic} contains the Student's 
 T-statistic of each regression coefficient and (4) \code{pValue} contains 
 the significance probability of each regression coefficient.} 
 \item{ResLevInf}{A data frame composed of three variables for each site in 
 the regression.  \code{Residual} contains the model residuals. 
 \code{Leverage} contains the leverage of each site.  \code{Influence} 
 contains the influence of each site.} \item{LevLim}{The critical value of 
 leverage.  See \code{\link{Leverage}}} \item{InflLim}{The critical value of 
 influence.  See \code{\link{Influence}}} \item{LevInf.Sig}{A logical matrix 
 indicating if the leverage (column 1) is significant and the influence 
 (column 2) is significant for each site in the regression.} 
 \item{PerformanceMetrics}{A list of not more than ten elements.  All 
 regression types return the mean squared error of residuals (\code{MSE}), 
 the coefficient of determination (\code{R2}), the adjusted coefficient of 
 determination (\code{R2_adj}) and the root mean squared error (\code{RMSE}, 
 in percent).  If \code{customWeight} contains model error variances, then
 the pseudo coefficient of regression (\code{R2_pseudo}), the average
 variance of prediction (\code{AVP}), the standard error of prediction
 (\code{Sp}, in percent), a vector of the individual variances of prediction
 for each site (\code{VP.PredVar}), the model-error variance
 (\code{ModErrVar}) and the standardized model error variance
 (\code{StanModErr}, in percent) are also returned.  Details on the
 appropriateness and applicability of performance metrics can be found in the
 WREG manual.} \item{X}{The input predictors.} \item{Y}{The input 
 observations.} \item{fitted.values}{A vector of model estimates from the 
 regression model.} \item{residuals}{A vector of model residuals.} 
 \item{Weighting}{The weighting matrix used to develop regression estimates.}
 \item{Input}{A list of input parameters for error searching.  Currently 
 empty.}
}
\description{
The \code{WREG.UW} function executes the multiple linear 
 regression analysis using a user-provided weighting matrix.
}
\details{
This function allows users to develop weights outside of the WREG 
 program and observe the resultant regressions.  Note that the weighting
 matrix must be invertible.
}
\examples{
# Import some example data
rm(list = ls())
peakFQdir <- paste0(
  file.path(system.file("exampleDirectory", package = "WREG"),
    "pfqImport"))
gisFilePath <- file.path(peakFQdir, "pfqSiteInfo.txt")
importedData <- importPeakFQ(pfqPath = peakFQdir, gisFile = gisFilePath)

# Organizing input data
Y <- importedData$Y$AEP_0.5
X <- importedData$X[c("Sand", "OutletElev", "Slope")]
transY <- "none"

# Make simple weighting using inverse record lengths
inverseRecLen <- diag(1 / diag(importedData$recLen))

# Run user-weights regression
result <- WREG.UW(Y, X, customWeight = inverseRecLen, transY)

}

