% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcorrPlot.R
\name{xcorPlot}
\alias{xcorPlot}
\title{Exploring correlation function}
\usage{
xcorPlot(object, alpha, theta, concurrentMin, DistMeth = 2, plot = TRUE)
}
\arguments{
\item{object}{A list of the imported data from \code{\link{importPeakFQ}} 
or\code{\link{importWREG}}.}

\item{alpha}{A number, \code{alpha} is a parameter used in the estimated 
cross-correlation between site records.  See equation 20 in the WREG v. 1.05
manual.}

\item{theta}{A number, \code{theta} is a parameter used in the estimated 
cross-correlation between site records.  See equation 20 in the WREG v. 1.05
manual.}

\item{concurrentMin}{A number specifying the minimum number of years of 
concurrent record required to estimate cross-correlation.}

\item{DistMeth}{A value of \code{1} indicates that the "Nautical Mile" 
approximation should be used to calculate inter-site distances.  A value of 
\code{2} designates the Haversine approximation.  See 
\code{\link{Dist.WREG}}.}

\item{plot}{A logical spcifying if the plot should be created.}
}
\value{
If \code{plot=FALSE}, the Nash-Sutcliffe model efficieny is returned.
}
\description{
\code{xcorPlot} explores the fit of the proposed approximation of
 cross-correlation.
}
\examples{
# Import some example data
peakFQdir <- paste0(
  file.path(system.file("exampleDirectory", package = "WREG"),
    "pfqImport"))
gisFilePath <- file.path(peakFQdir, "pfqSiteInfo.txt")
importedData <- importPeakFQ(pfqPath = peakFQdir, gisFile = gisFilePath)

# Build cross-correlation plot
xcorPlot(object = importedData, alpha = 0.01, theta = 0.98,
  concurrentMin = 10)
}

