% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importWREG.R
\name{importWREG}
\alias{importWREG}
\title{Import Data from Old WREG Format}
\usage{
importWREG(wregPath, sites = "")
}
\arguments{
\item{wregPath}{A directory that contains all of the files needed to 
implement the MatLab version of WREG.}

\item{sites}{(optional) A vetor of sites that should be return.  Allows for 
data subsetting.}
}
\value{
All outputs are returned as part of a list.  The list includes:
\item{sites}{A vector of site IDs.}
\item{Y}{A data frame whose comlumns represent unique frequency events, 
while the row represent particular sites in the same order as \code{sites}.}
\item{X}{A data frame whose columns represent basin characteristics to be used
 as dependent variables and whose rows represent sites corresponding 
 to \code{sites}.}
\item{LP3f}{A matrix containing the fitted LP# parameters that are fixed 
across exceedence probability.  These include the standard deviation, 
skew and regional skew for each site.}
\item{LP3k}{A matrix of the fitted kappa parameters of the LP3 distribution
 for each \code{AEP}.}
\item{BasChars}{A matrix containing the site IDs, latitudes and longitudes.}
\item{recLen}{A square matrix indicating the number of overlapping years
 for each site pair.}
\item{recCor}{A matrix of the correlaiton between site paris.}
\item{UW}{A matrix of user weights, if included.}
}
\description{
The \code{importWREG} function reads the WREG inputs from a directory 
set up for the old WREG program.
}
\details{
This functions allows users to use the legacy format of WREG.  This includes
an established directory that contains valid \dQuote{SiteInfo.txt}, 
\dQuote{FlowChar.txt}, \dQuote{LP3G.txt}, \dQuote{LP3K.txt}, 
\dQuote{LP3s.txt} and \dQuote{USGS##########.txt}, multiple files 
containing time series for each site.  The file \dQuote{UserWLS.txt} is 
optional.  For further information on the format of these files, see the 
program manual (Techniques and Methods 4-A8).  Files that are not valid 
inputs files for the WREG version describe therein will not be accepted.
}

