% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega.WLS.ROImatchMatLab.R
\name{Omega.WLS.ROImatchMatLab}
\alias{Omega.WLS.ROImatchMatLab}
\title{Weighing Matrix for ROI-WLS (WREG)}
\usage{
Omega.WLS.ROImatchMatLab(Y.all, X.all, LP3.all, RecordLengths.all, NDX)
}
\arguments{
\item{Y.all}{The dependent variable of interest at all sites in the network,
with any transformations already applied.}

\item{X.all}{The independent variables  at all sites in the network, with any 
transformations already applied.  Each row represents a site and each column
represents a particular independent variable.  The rows must be in the same
order as the dependent variables in \code{Y}.}

\item{LP3.all}{A dataframe containing the fitted Log-Pearson Type III standard 
deviate, standard deviation and skew for all sites in the network.  The
names of this data frame are \code{S}, \code{K} and \code{G}.  The order of
the rows must be the same as \code{Y}.}

\item{RecordLengths.all}{\code{RecordLengths.all} may be a matrix whose rows
and columns are in the same order as \code{Y}.  Each \code{(r,c)} element
represents the length of concurrent record between sites \code{r} and
\code{c}.  The diagonal elements therefore represent each site's full record
length.  For \dQuote{WLS}, only the at-site record lengths are needed.  In
this case, \code{RecordLengths} can be a vector or a matrix.  Here, this
should only include sites in the current region of influence.}

\item{NDX}{A vector listing the indices of the sites that comprise the region
of influence.}
}
\value{
\code{Omega.WLS.ROImatchMatLab} returns a list with three elements: 
 \item{Omega}{The estimated weighting matrix.} \item{var.modelerror.0}{The
 estimated model error variance for a constant-value model.} 
 \item{var.modelerror.k}{The estimated model error variance for a k-variable
 model.}
}
\description{
\code{Omega.WLS.ROImatchMatLab} calculates the weighting function
for a weighted least-squares regression using regions-of-influence.  This is
largely legacy code to match WREG v. 1.05 idiosyncrasies.
}
\details{
This is a legacy function that matches the idiosyncrasies of WREG v.
 1.05. This includes using all sites to implement the sigma regression,
 averaging across all record lengths, using arbitrary record lengths to
 estimate weights and using a new function for estimating the MSE of the
 basic OLS model.
 
 This function will become obsolete once all idiosyncrasies are assessed.
}
\examples{
# Import some example data
peakFQdir <- paste0(
  file.path(system.file("exampleDirectory", package = "WREG"),
    "pfqImport"))
gisFilePath <- file.path(peakFQdir, "pfqSiteInfo.txt")
importedData <- importPeakFQ(pfqPath = peakFQdir, gisFile = gisFilePath)

# Organizing input data
lp3Data <- importedData$LP3f
lp3Data$K <- importedData$LP3k$AEP_0.5
Y <- importedData$Y$AEP_0.5
X <- importedData$X[c("Sand", "OutletElev", "Slope")]

#### Geographic Region-of-Influence
i <- 1 # Site of interest
n <- 10 # size of region of influence
Gdist <- vector(length=length(Y)) # Empty vector for geographic distances
for (j in 1:length(Y)) {
  if (i!=j) {
    #### Geographic distance
    Gdist[j] <- Dist.WREG(Lat1 = importedData$BasChars$Lat[i],
      Long1 = importedData$BasChars$Long[i],
      Lat2 = importedData$BasChars$Lat[j],
      Long2 = importedData$BasChars$Long[j]) # Intersite distance, miles
  } else {
    Gdist[j] <- Inf # To block self identification.
  }
}
temp <- sort.int(Gdist,index.return=TRUE)
NDX <- temp$ix[1:n] # Sites to use in this regression

# Compute weighting matrix
weightingResult <- Omega.WLS.ROImatchMatLab(Y.all = Y, X.all = X,
  LP3.all = lp3Data, RecordLengths.all = importedData$recLen, NDX = NDX)

}
